+++ mod_editor/apf_studio/__init__.py
@@ -12,7 +12,7 @@ from .models import (
     UniformAsset,
 )
 
-__version__ = "0.1.0-alpha.38"
+__version__ = "0.1.0-alpha.39"
 
 __all__ = [
     "ApfAsset",